Complete initial build
Remove the renderer foundations from last night.
Switch to a class-based system where "entities" are created.
Each entity is given a starting position and velocity when instantiated.
Each entity then handles updating it's own position, drawing itself to
the canvas and handling collisions with other entities and the
canvas "walls".

A "world" class can then be created to hold all entities and to
handle drawing the entire "world" on each frame, and asking the
entities to check for collisions.

For now, we're creating a default "world" of 100 random entities when
the app starts up.
Next will  be to add some actual UI which allows starting,
stopping/pausing, spawning entities, controlling velcoity range.